+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to
   [#3783](https://github.com/OpenFn/lightning/issues/3783)
 - Enforce readonly state in collaborative editor forms for viewers and old
   snapshots [#3948](https://github.com/OpenFn/lightning/pull/3948)
+- Collab Editor: Add Workflow YAML code viewer panel
+  [#3646](https://github.com/OpenFn/lightning/issues/3646)
 - Webhook authentication management in the collaborative editor
   [#3887](https://github.com/OpenFn/lightning/issues/3887)
 - Mix task to merge project state files without database access